Add blocks_and_iterators.rb and update newline in everything_is_an_object.rb
This commit introduces a new file, blocks_and_iterators.rb, with a few simple demonstrations of block syntax and iterators in Ruby. It also includes a minor correction in everything_is_an_object.rb to conform to stylistic standards, specifically by adding a newline at the end of the file.